Update 1491. Average Salary Excluding the Minimum and Maximum Salary.cpp

+++ Codes/1491. Average Salary Excluding the Minimum and Maximum Salary.cpp
@@ -2,12 +2,12 @@ class Solution {
 public:
     double average(vector<int>& salary) {
         int n = salary.size()-2;
-        int sum = 0;
+        double sum = 0;
         sort(salary.begin(),salary.end());
         for(int i=1;i<salary.size()-1;i++){
             sum = sum + salary[i];
         }
-        int avg = sum/n;
+        double avg = sum/n;
         return avg;
     }
 };
